prep/VECTOR-DATABASE/pyproject.toml
[project]
name = "vector-database"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "qdrant-client>=1.14.3",
]
